Removed stray size_t cast that was causing a warning in avtConnComponentsExpression
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@26513 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Expressions/General/avtConnComponentsExpression.C
@@ -3236,7 +3236,7 @@ avtConnComponentsExpression::SpatialPartition::CreatePartition
                 if (b_list[i]->IsDone())
                     continue;
                 int nregions = b_list[i]->GetNumberOfRegions();
-                for (int j = 0 ; j < (size_t)nregions ; j++)
+                for (int j = 0 ; j < nregions ; j++)
                 {
                     b_list[i]->SetNumberOfPointsForRegion(j, 0);
                 }
